+++ Content.Server/_NF/GameRule/NfAdventureRuleSystem.cs
@@ -8,16 +8,13 @@ using System.Threading.Tasks;
 using Content.Shared._NF.GameRule;
 using Content.Server._NF.GameTicking.Events;
 using Robust.Server.GameObjects;
-using Robust.Server.Maps;
 using Content.Shared.GameTicking.Components;
 using Robust.Shared.Map;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Random;
-using Content.Server.Shuttles.Systems;
 using Content.Server.Cargo.Components;
 using Content.Server.GameTicking;
 using Content.Server.GameTicking.Rules;
-using Content.Server.Maps;
 using Content.Server.Station.Systems;
 using Content.Shared._NF.CCVar; // Frontier
 using Robust.Shared.Configuration;
@@ -29,14 +26,13 @@ using Robust.Shared.Network;
 using Content.Shared.GameTicking;
 using Robust.Shared.Enums;
 using Robust.Server.Player;
-using Content.Server.Warps;
 
 namespace Content.Server._NF.GameRule;
 
 /// <summary>
 /// This handles the dungeon and trading post spawning, as well as round end capitalism summary
 /// </summary>
-public sealed class NfAdventureRuleSystem : GameRuleSystem<AdventureRuleComponent>
+public sealed class NFAdventureRuleSystem : GameRuleSystem<AdventureRuleComponent>
 {
     [Dependency] private readonly IPrototypeManager _prototypeManager = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
@@ -45,10 +41,9 @@ public sealed class NfAdventureRuleSystem : GameRuleSystem<AdventureRuleComponen
     [Dependency] private readonly MapLoaderSystem _map = default!;
     [Dependency] private readonly MetaDataSystem _meta = default!;
     [Dependency] private readonly StationSystem _station = default!;
-    [Dependency] private readonly ShuttleSystem _shuttle = default!;
-    [Dependency] private readonly PhysicsSystem _physics = default!;
     [Dependency] private readonly BankSystem _bank = default!;
     [Dependency] private readonly StationRenameWarpsSystems _renameWarps = default!;
+    [Dependency] private readonly PointOfInterestSystem _poi = default!;
 
     private readonly HttpClient _httpClient = new();
 
@@ -237,233 +232,16 @@ public sealed class NfAdventureRuleSystem : GameRuleSystem<AdventureRuleComponen
                 remainingUniqueProtosBySpawnGroup[location.SpawnGroup].Add(location);
             }
         }
-        GenerateDepots(depotProtos, out component.CargoDepots);
-        GenerateMarkets(marketProtos, out component.MarketStations);
-        GenerateRequireds(requiredProtos, out component.RequiredPois);
-        GenerateOptionals(optionalProtos, out component.OptionalPois);
-        GenerateUniques(remainingUniqueProtosBySpawnGroup, out component.UniquePois);
+        _poi.GenerateDepots(depotProtos, out component.CargoDepots);
+        _poi.GenerateMarkets(marketProtos, out component.MarketStations);
+        _poi.GenerateRequireds(requiredProtos, out component.RequiredPois);
+        _poi.GenerateOptionals(optionalProtos, out component.OptionalPois);
+        _poi.GenerateUniques(remainingUniqueProtosBySpawnGroup, out component.UniquePois);
 
         base.Started(uid, component, gameRule, args);
 
         // Using invalid entity, we don't have a relevant entity to reference here.
         RaiseLocalEvent(EntityUid.Invalid, new StationsGeneratedEvent(), broadcast: true); // TODO: attach this to a meaningful entity.
-    }
-
-    private void GenerateDepots(List<PointOfInterestPrototype> depotPrototypes, out List<EntityUid> depotStations)
-    {
-        //For depots, we want them to fill a circular type dystance formula to try to keep them as far apart as possible
-        //Therefore, we will be taking our range properties and treating them as magnitudes of a direction vector divided
-        //by the number of depots set in our corresponding cvar
-
-        depotStations = new List<EntityUid>();
-        var depotCount = _configurationManager.GetCVar(NFCCVars.CargoDepots);
-        var rotation = 2 * Math.PI / depotCount;
-        var rotationOffset = _random.NextAngle() / depotCount;
-
-        for (int i = 0; i < depotCount && depotPrototypes.Count > 0; i++)
-        {
-            var proto = _random.Pick(depotPrototypes);
-            Vector2i offset = new Vector2i((int) (_random.Next(proto.MinimumDistance, proto.MaximumDistance) * _distanceOffset), 0);
-            offset = offset.Rotate(rotationOffset);
-            rotationOffset += rotation;
-            // Append letter to depot name.
-
-            string overrideName = proto.Name;
-            if (i < 26)
-                overrideName += $" {(char) ('A' + i)}"; // " A" ... " Z"
-            else
-                overrideName += $" {i + 1}"; // " 27", " 28"...
-            if (TrySpawnPoiGrid(proto, offset, out var depotUid, overrideName: overrideName) && depotUid is { Valid: true } depot)
-            {
-                depotStations.Add(depot);
-                AddStationCoordsToSet(offset); // adjust list of actual station coords
-            }
-        }
-    }
-
-    private void GenerateMarkets(List<PointOfInterestPrototype> marketPrototypes, out List<EntityUid> marketStations)
-    {
-        //For market stations, we are going to allow for a bit of randomness and a different offset configuration. We dont
-        //want copies of this one, since these can be more themed and duplicate names, for instance, can make for a less
-        //ideal world
-
-        marketStations = new List<EntityUid>();
-        var marketCount = _configurationManager.GetCVar(NFCCVars.MarketStations);
-        _random.Shuffle(marketPrototypes);
-        int marketsAdded = 0;
-        foreach (var proto in marketPrototypes)
-        {
-            if (marketsAdded >= marketCount)
-                break;
-
-            var offset = GetRandomPOICoord(proto.MinimumDistance, proto.MaximumDistance, true);
-
-            if (TrySpawnPoiGrid(proto, offset, out var marketUid) && marketUid is { Valid: true } market)
-            {
-                marketStations.Add(market);
-                marketsAdded++;
-                AddStationCoordsToSet(offset);
-            }
-        }
-    }
-
-    private void GenerateOptionals(List<PointOfInterestPrototype> optionalPrototypes, out List<EntityUid> optionalStations)
-    {
-        //Stations that do not have a defined grouping in their prototype get a default of "Optional" and get put into the
-        //generic random rotation of POIs. This should include traditional places like Tinnia's rest, the Science Lab, The Pit,
-        //and most RP places. This will essentially put them all into a pool to pull from, and still does not use the RNG function.
-
-        optionalStations = new List<EntityUid>();
-        var optionalCount = _configurationManager.GetCVar(NFCCVars.OptionalStations);
-        _random.Shuffle(optionalPrototypes);
-        int optionalsAdded = 0;
-        foreach (var proto in optionalPrototypes)
-        {
-            if (optionalsAdded >= optionalCount)
-                break;
-
-            var offset = GetRandomPOICoord(proto.MinimumDistance, proto.MaximumDistance, true);
-
-            if (TrySpawnPoiGrid(proto, offset, out var optionalUid) && optionalUid is { Valid: true } uid)
-            {
-                optionalStations.Add(uid);
-                AddStationCoordsToSet(offset);
-            }
-        }
-    }
-
-    private void GenerateRequireds(List<PointOfInterestPrototype> requiredPrototypes, out List<EntityUid> requiredStations)
-    {
-        //Stations are required are ones that are vital to function but otherwise still follow a generic random spawn logic
-        //Traditionally these would be stations like Expedition Lodge, NFSD station, Prison/Courthouse POI, etc.
-        //There are no limit to these, and any prototype marked alwaysSpawn = true will get pulled out of any list that isnt Markets/Depots
-        //And will always appear every time, and also will not be included in other optional/dynamic lists
-
-        requiredStations = new List<EntityUid>();
-        foreach (var proto in requiredPrototypes)
-        {
-            var offset = GetRandomPOICoord(proto.MinimumDistance, proto.MaximumDistance, true);
-
-            if (TrySpawnPoiGrid(proto, offset, out var requiredUid) && requiredUid is { Valid: true } uid)
-            {
-                requiredStations.Add(uid);
-                AddStationCoordsToSet(offset);
-            }
-        }
-    }
-
-    private void GenerateUniques(Dictionary<string, List<PointOfInterestPrototype>> uniquePrototypes, out List<EntityUid> uniqueStations)
-    {
-        //Unique locations are semi-dynamic groupings of POIs that rely each independantly on the SpawnChance per POI prototype
-        //Since these are the remainder, and logically must have custom-designated groupings, we can then know to subdivide
-        //our random pool into these found groups.
-        //To do this with an equal distribution on a per-POI, per-round percentage basis, we are going to ensure a random
-        //pick order of which we analyze our weighted chances to spawn, and if successful, remove every entry of that group
-        //entirely.
-
-        uniqueStations = new List<EntityUid>();
-        foreach (var prototypeList in uniquePrototypes.Values)
-        {
-            // Try to spawn
-            _random.Shuffle(prototypeList);
-            foreach (var proto in prototypeList)
-            {
-                var chance = _random.NextFloat(0, 1);
-                if (chance <= proto.SpawnChance)
-                {
-                    var offset = GetRandomPOICoord(proto.MinimumDistance, proto.MaximumDistance, true);
-
-                    if (TrySpawnPoiGrid(proto, offset, out var optionalUid) && optionalUid is { Valid: true } uid)
-                    {
-                        uniqueStations.Add(uid);
-                        AddStationCoordsToSet(offset);
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    private bool TrySpawnPoiGrid(PointOfInterestPrototype proto, Vector2 offset, out EntityUid? gridUid, string? overrideName = null)
-    {
-        gridUid = null;
-        if (_map.TryLoad(_mapId, proto.GridPath.ToString(), out var mapUids,
-                new MapLoadOptions
-                {
-                    Offset = offset,
-                    Rotation = _random.NextAngle()
-                }))
-        {
-
-            string stationName = string.IsNullOrEmpty(overrideName) ? proto.Name : overrideName;
-
-            EntityUid? stationUid = null;
-            if (_prototypeManager.TryIndex<GameMapPrototype>(proto.ID, out var stationProto))
-            {
-                stationUid = _station.InitializeNewStation(stationProto.Stations[proto.ID], mapUids, stationName);
-            }
-
-            foreach (var grid in mapUids)
-            {
-                var meta = EnsureComp<MetaDataComponent>(grid);
-                _meta.SetEntityName(grid, stationName, meta);
-
-                EntityManager.AddComponents(grid, proto.AddComponents);
-            }
-
-            // Rename warp points after set up if needed
-            if (proto.NameWarp)
-            {
-                bool? hideWarp = proto.HideWarp ? true : null;
-                if (stationUid != null)
-                    _renameWarps.SyncWarpPointsToStation(stationUid.Value, forceAdminOnly: hideWarp);
-                else
-                    _renameWarps.SyncWarpPointsToGrids(mapUids, forceAdminOnly: hideWarp);
-            }
-
-            gridUid = mapUids[0];
-            return true;
-        }
-
-        return false;
-    }
-
-    private Vector2 GetRandomPOICoord(float unscaledMinRange, float unscaledMaxRange, bool scaleRange)
-    {
-        int numRetries = int.Max(_configurationManager.GetCVar(NFCCVars.POIPlacementRetries), 0);
-        float minDistance = float.Max(_configurationManager.GetCVar(NFCCVars.MinPOIDistance), 0); // Constant at the end to avoid NaN weirdness
-
-        Vector2 coords = _random.NextVector2(unscaledMinRange, unscaledMaxRange);
-        if (scaleRange)
-            coords *= _distanceOffset;
-        for (int i = 0; i < numRetries; i++)
-        {
-            bool positionIsValid = true;
-            foreach (var station in _stationCoords)
-            {
-                if (Vector2.Distance(station, coords) < minDistance)
-                {
-                    positionIsValid = false;
-                    break;
-                }
-            }
-
-            // We have a valid position
-            if (positionIsValid)
-                break;
-
-            // No vector yet, get next value.
-            coords = _random.NextVector2(unscaledMinRange, unscaledMaxRange);
-            if (scaleRange)
-                coords *= _distanceOffset;
-        }
-
-        return coords;
-    }
-
-    private void AddStationCoordsToSet(Vector2 coords)
-    {
-        _stationCoords.Add(coords);
     }
 
     private async Task ReportRound(string message, int color = 0x77DDE7)
